backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.5.3",
    "django==5.2",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "faker>=37.11.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
]
